fix: rebind Mod4+e to open emacs instead of neovim
Yes. It has come to this. In such a short time, indeed. I'm sorry it had
to happen like this, and for you to find out in such a way within this
mess of commit logs. I'll still use you, you're still here, and I'll
make sure that the jobs you're used for will be more appropriate. I was
asking too much of you, forcing you to turn into something you're not, a
fully fledged IDE.
Sorry, Neovim. x